+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - [#182](https://github.com/Datatamer/unify-client-python/issues/182) Add the ability to refresh estimated pair counts.
   - [#184](https://github.com/Datatamer/unify-client-python/issues/184) Support for getting published cluster configurations
   - [#201](https://github.com/Datatamer/unify-client-python/issues/201) Support for refreshing published cluster IDs
+  - [#112](https://github.com/Datatamer/unify-client-python/issues/112) Support for attribute configurations
 
 ## 0.7.0
   **BREAKING CHANGES**